+++ src/pywmapi/common/__init__.py
@@ -1,5 +1,7 @@
-from .models import ModelBase
+from .defs import *
 from .enums import *
+from .models import *
 
 HOMEPAGE_URL = "https://warframe.market"
 API_BASE_URL = "https://api.warframe.market/v1"
+WSS_BASE_URL = "wss://warframe.market/socket"