+++ src/pywmapi/common/defs.py
@@ -1,0 +1,3 @@
+from typing import Callable
+
+MessageCallback = Callable[[str], None]