+++ src/pywmapi/items/models.py
@@ -1,20 +1,16 @@
 from dataclasses import dataclass
 from datetime import datetime
 from enum import Enum
-from typing import Dict, Any, Optional, List
+from typing import Any, Dict, List, Optional
+
 from dacite import Config
 
 from ..common import *
 from ..lang.models import LangInItem
-from ..auth.models import UserShort
-
 
 __all__ = [
     "ItemShort",
     "ItemFull",
-    "OrderCommon",
-    "OrderRow",
-    # "OrderFull",
 ]
 
 
@@ -28,8 +24,7 @@ class ItemShort(ModelBase):
 
 @dataclass
 class ItemFull(ModelBase):
-    @dataclass
-    class Rarity(Enum):
+    class Rarity(str, Enum):
         common = "common"
         uncommon = "uncommon"
         rare = "rare"
@@ -65,28 +60,3 @@ class ItemFull(ModelBase):
     es: Optional[LangInItem]
     it: Optional[LangInItem]
     pl: Optional[LangInItem]
-
-
-@dataclass
-class OrderCommon(ModelBase):
-    id: str
-    platinum: int
-    quantity: int
-    order_type: OrderType
-    platform: Platform
-    region: Optional[str]
-    creation_date: datetime
-    last_update: datetime
-    visible: bool
-
-
-@dataclass
-class OrderRow(OrderCommon):
-    user: UserShort
-
-
-@dataclass
-class OrderFull(OrderRow):
-    # TODO
-    def __init__(self):
-        raise NotImplementedError()