+++ test/test_rivens_api.py
@@ -1,0 +1,10 @@
+from pywmapi.common import *
+from pywmapi.rivens import *
+
+
+def test_list_items():
+    list_items()
+
+
+def test_list_attrs():
+    list_attrs()